Extract RedirectNodes module; start test module
